requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.29,<1
requests>=2.31,<3
openpyxl>=3.1,<4
SQLAlchemy>=2.0,<3
alembic>=1.13,<2
PyMySQL>=1.1,<2
python-dotenv>=1.0,<2
pandas>=2.2,<3
python-multipart>=0.0.9,<1
PyMuPDF>=1.24,<2
python-pptx>=1.0,<2
beautifulsoup4>=4.12,<5
browser-cookie3>=0.20,<1
DrissionPage>=4.1,<5
